decode signed timestamps correctly for AFP 2.x sessions
AFP date fields are signed seconds relative to 2000-01-01. Decode them before applying the epoch and retain the AFP 2.x server-to-client clock offset from FPGetSrvrParms. Use the unset-date sentinel only as a directory fallback, preserving real timestamps.

Widen public and IPC date fields to retain valid pre-2000 values and add protocol coverage for signed values, the offset, and the sentinel.

+++ include/netatalk-client/afpsl.h
@@ -47,8 +47,8 @@ struct afp_sl_desktop_appl {
     uint32_t tag;
     uint32_t directory_id;
     uint32_t file_id;
-    uint32_t creation_date;
-    uint32_t modification_date;
+    int64_t creation_date;
+    int64_t modification_date;
     uint64_t data_fork_size;
     uint64_t resource_fork_size;
     char pathname[AFPC_MAX_NAME_BYTES];
